src/app/components/project-files/requirements.txt
python-nmap==0.7.1
Flask==3.0.0
reportlab==4.0.7
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
